refactor: extract db-worker task dispatcher into testable module
Move PanelAwareFilter, resolvePanelIntervalsInPlace, and the 25+ case
switch dispatcher from db-worker.ts into db-worker-dispatch.ts with
explicit dependency injection (DispatchDependencies). db-worker.ts now
delegates to dispatchTask(). Add 14 tests covering task dispatch and
panel filter cleanup.
